[APM][Refactor] Rename create_assets_es_client to create_entities_es_clients
