+++ x-pack/plugins/observability_solution/apm/server/routes/custom_dashboards/route.ts
@@ -14,7 +14,7 @@ import { getCustomDashboards } from './get_custom_dashboards';
 import { getServicesWithDashboards } from './get_services_with_dashboards';
 import { getApmEventClient } from '../../lib/helpers/get_apm_event_client';
 import { rangeRt } from '../default_api_types';
-import { createEntitiesESClient } from '../../lib/helpers/create_es_client/create_assets_es_client/create_assets_es_clients';
+import { createEntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
 import { getEntitiesWithDashboards } from './get_entities_with_dashboards';
 
 const serviceDashboardSaveRoute = createApmServerRoute({
