+++ x-pack/plugins/observability_solution/apm/server/routes/entities/get_service_entities_history_metrics.ts
@@ -16,7 +16,7 @@ import {
   ENTITY_METRICS_THROUGHPUT,
   LAST_SEEN,
 } from '../../../common/es_fields/entities';
-import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_assets_es_client/create_assets_es_clients';
+import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
 
 interface Params {
   entitiesESClient: EntitiesESClient;
