+++ x-pack/plugins/observability_solution/apm/server/routes/entities/get_service_latest_entity.ts
@@ -14,7 +14,7 @@ import {
 import { ENTITY, ENTITY_TYPE } from '../../../common/es_fields/entities';
 import { environmentQuery } from '../../../common/utils/environment_query';
 import { isFiniteNumber } from '../../../common/utils/is_finite_number';
-import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_assets_es_client/create_assets_es_clients';
+import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
 import { entitiesRangeQuery } from './get_entities';
 import { EntitiesRaw, EntityType, ServiceEntities } from './types';
