+++ x-pack/plugins/observability_solution/apm/server/routes/entities/services/routes.ts
@@ -9,7 +9,7 @@ import { jsonRt } from '@kbn/io-ts-utils';
 import * as t from 'io-ts';
 import { EntityServiceListItem } from '../../../../common/entities/types';
 import { environmentQuery } from '../../../../common/utils/environment_query';
-import { createEntitiesESClient } from '../../../lib/helpers/create_es_client/create_assets_es_client/create_assets_es_clients';
+import { createEntitiesESClient } from '../../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
 import { createApmServerRoute } from '../../apm_routes/create_apm_server_route';
 import { environmentRt, kueryRt, rangeRt } from '../../default_api_types';
 import { getServiceEntities } from './get_service_entities';
