+++ x-pack/plugins/observability_solution/apm/server/routes/historical_data/has_historical_entities_data.ts
@@ -7,7 +7,7 @@
 import { WrappedElasticsearchClientError } from '@kbn/observability-plugin/server';
 import { Logger } from '@kbn/core/server';
 import { errors } from '@elastic/elasticsearch';
-import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_assets_es_client/create_assets_es_clients';
+import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
 
 export async function hasEntitiesData(entitiesESClient: EntitiesESClient, logger: Logger) {
   const params = {
